+++ packages/ecma402-abstract/DateTimeFormat/tsconfig.json
@@ -1,0 +1,17 @@
+{
+  "compilerOptions": {
+    "composite": true,
+    "paths": {
+      "#packages/*": ["../../../packages/*"]
+    }
+  },
+  "extends": "../../../tsconfig.json",
+  "references": [
+    {
+      "path": "../tsconfig.json"
+    },
+    {
+      "path": "../types/tsconfig.json"
+    }
+  ]
+}
