+++ packages/intl-datetimeformat/src/core.ts
@@ -12,13 +12,13 @@ import {
 } from '#packages/ecma402-abstract/types/date-time.js'
 import {defineProperty, invariant} from '#packages/ecma402-abstract/utils.js'
 import Decimal from '@formatjs/bigdecimal'
-import {FormatDateTime} from './abstract/FormatDateTime.js'
-import {FormatDateTimeRange} from './abstract/FormatDateTimeRange.js'
-import {FormatDateTimeRangeToParts} from './abstract/FormatDateTimeRangeToParts.js'
-import {FormatDateTimeToParts} from './abstract/FormatDateTimeToParts.js'
-import {InitializeDateTimeFormat} from './abstract/InitializeDateTimeFormat.js'
-import {parseDateTimeSkeleton} from './abstract/skeleton.js'
-import {DATE_TIME_PROPS} from './abstract/utils.js'
+import {FormatDateTime} from '#packages/ecma402-abstract/DateTimeFormat/FormatDateTime.js'
+import {FormatDateTimeRange} from '#packages/ecma402-abstract/DateTimeFormat/FormatDateTimeRange.js'
+import {FormatDateTimeRangeToParts} from '#packages/ecma402-abstract/DateTimeFormat/FormatDateTimeRangeToParts.js'
+import {FormatDateTimeToParts} from '#packages/ecma402-abstract/DateTimeFormat/FormatDateTimeToParts.js'
+import {InitializeDateTimeFormat} from '#packages/ecma402-abstract/DateTimeFormat/InitializeDateTimeFormat.js'
+import {parseDateTimeSkeleton} from '#packages/ecma402-abstract/DateTimeFormat/skeleton.js'
+import {DATE_TIME_PROPS} from '#packages/ecma402-abstract/DateTimeFormat/utils.js'
 import links from './data/links.generated.js'
 import getInternalSlots from './get_internal_slots.js'
 import {unpack} from './packer.js'
