+++ packages/intl-datetimeformat/tests/abstract/skeleton.test.ts
@@ -2,7 +2,7 @@ import {
   parseDateTimeSkeleton,
   processDateTimePattern,
   splitRangePattern,
-} from '../../src/abstract/skeleton'
+} from '#packages/ecma402-abstract/DateTimeFormat/skeleton.js'
 import {expect, test} from 'vitest'
 test('splitRangePattern basic case', function () {
   expect(splitRangePattern('{month} {day} - {day}')).toEqual([
